+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.1.0-dev.1
 
 - The initial `draft` version.
+
+## 0.1.0
+
+- Added the `byte-length` official SHA512 test vectors.
+- Added `hashing` example
+- Allowed `personalisation` paramater to be less then 16 bytes long by zero-padding to 16 bytes.